chore: rename to EmptyResults
