+++ src/SearchInput/components/EmptyResults.tsx
